==> notebooks/TP2_NER/06_results_and_comparison.ipynb <==
## Imports & configuration

import pandas as pd


## Medical domain

results_medical = pd.DataFrame([
    {
        "Model": "LSTM (Word2Vec)",
        "Precision": 0.8475,
        "Recall": 1.0000,
        "F1-score": 0.9174,
        "Accuracy": 0.8475
    },
    {
        "Model": "CNN (FastText)",
        "Precision": 0.8587,
        "Recall": 0.9945,
        "F1-score": 0.9217,
        "Accuracy": 0.8567
    },
    {
        "Model": "CamemBERT",
        "Precision": 0.8910,
        "Recall": 0.9800,
        "F1-score": 0.9424,
        "Accuracy": 0.8910
    }
])

results_medical


## Press domain

results_press = pd.DataFrame([
    {
        "Model": "LSTM (Word2Vec)",
        "Precision": 0.8755,
        "Recall": 0.8873,
        "F1-score": 0.8814,
        "Accuracy": 0.8251
    },
    {
        "Model": "CNN (FastText)",
        "Precision": 0.8637,
        "Recall": 0.8888,
        "F1-score": 0.8761,
        "Accuracy": 0.8159
    },
    {
        "Model": "CamemBERT",
        "Precision": 0.8871,
        "Recall": 0.9679,
        "F1-score": 0.9258,
        "Accuracy": 0.9160
    }
])

results_press


results_medical["Domain"] = "Medical"
results_press["Domain"] = "Press"

results_all = pd.concat([results_medical, results_press], ignore_index=True)
results_all
